PP-1 Hilt added

+++ app/src/main/java/com/empty/picpap/core/data/FavoritePhotoRepository.kt
@@ -6,5 +6,5 @@ interface FavoritePhotoRepository {
 
     suspend fun getFavoritePhotos(): DataResult<List<PhotoData>>
 
-    fun removePhoto(id: Int)
+    fun removePhoto(id: Long)
 }
